Create Problem 3 file
Problem 3
---------

In Laxmitown, one can exchange a coin of amount X, with three coins of amounts X/2, X/3 and X/4.
But coins cannot be converted if the result is a fractional value. Only whole numbers are valid.

For example, if one deposits a coin worth 36, he get back 3 coins: 18, 12 and 9, with a value of 39
Further 12 can be exchanged for coins: 6, 4 and 3, totalling to 13. The final value being 39 + 13 = 52.
Note that 18 cannot be exchanged further, because 18/4 does not result in a whole number.

Given a number X, write a program to find the maximum value it can be exhanged for.

Owner -Sanjay